Cleaning file structures
Basic changes to structure are made. This will need to be diffed with the develop branch to ensure no good changes are overwritten.
Reasonably sure I tested all files.
