+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "firebase": "^5.9.0",
+    "js-cookie": "^2.2.0",
     "react": "^16.8.5",
     "react-dom": "^16.8.4",
     "react-scripts": "2.1.8"
