--- src/Button.css
@@ -1,14 +1,0 @@
-.Button {
-        padding: 20px;
-        border: none;
-        background-color: #ADD8E6;
-        text-align: center;
-        text-decoration: none;
-        display: inline-block;
-        border-radius: 12px;
-        transition-duration: 0.4s;
-        width: 20%;
-}
-.Button:hover{
-        background-color: #B0C4DE;
-}
