+++ src/components/NavBar.css
@@ -1,0 +1,100 @@
+.ui-header {
+    background-color: #111111;
+    padding: 10px;
+    padding-left: 25px;
+    display: inline-flex;
+    width: 100%;
+
+
+}
+/* Add a black background color to the top navigation bar */
+
+
+
+/* Style the links inside the navigation bar */
+.ui-header a {
+    /*float: left;*/
+    /*display: inline-list-item;*/
+    color: white;
+    /*text-align: center;*/
+    /*padding: 15px 15px;*/
+    /*text-decoration: none;*/
+    /*font-size: 17px;*/
+}
+
+
+
+
+/* Change the color of links on hover */
+.ui-header a:hover {
+    background-color: #515151;
+    color: #ffffff;
+}
+
+
+/* Style the "active" element to highlight the current page */
+
+.ui-header a.active {
+    /*background-color: #fb0208;*/
+    color: white;
+}
+
+
+/* ---------- Logo ---------- */
+.ui-header.logo {
+    height: 50px;
+}
+
+/*
+.ui-header.logo a:hover {
+    background-color: #111111;
+    color: black;
+}
+*/
+
+
+/* ---------- Home ---------- */
+.home {
+    /*background-color: blue;*/
+    margin-left: 15px;
+    margin-right: 15px;
+    align-items: center;
+    /*border-radius: 5px;*/
+    /*border-bottom-color: #fb0208;*/
+    /*border: #61dafb;*/
+}
+
+
+
+/* ---------- Search ---------- */
+.ui-header.search-bar {
+
+}
+
+.search-bar {
+    display: inline-block;
+    height: 30px;
+    margin-left: 200px;
+    min-width: 200px;
+    padding-left: 100px;
+}
+
+/* Style the search box inside the navigation bar */
+/*
+.ui-header input[type=text] {
+    float: right;
+    padding: 6px;
+    border: none;
+    margin-top: 16px;
+    margin-right: 16px;
+    font-size: 17px;
+    color: white;
+}
+*/
+
+/* ---------- Profile ---------- */
+.ui-header.profile {
+    display: inline-block;
+    margin-top: 160px;
+    float: right;
+}
