+++ src/index.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import HistoryProfile from './History_Profile';
 import App from './App';
-import HistoryPage from './History_Page';
-import * as serviceWorker from './serviceWorker';
-import MoviePage from "./Movie_Page";
-import SignIn from "./SignIn";
+import HistoryPage from './historyPage/History_Page';
+import * as serviceWorker from './libraries/serviceWorker';
+import MoviePage from "./moviePage/Movie_Page";
+import SignIn from "./loginPage/SignIn";
+
 
 //You can change this render function to check on your components
 //Change HistoryProfile to YourAppFile
 
-ReactDOM.render(<SignIn />, document.getElementById('root'));
+ReactDOM.render(<HistoryPage />, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
